Move steps frame into same module as main window
